18. 四数之和 (4Sum)

+++ LeetCodeText.java
@@ -200,26 +200,26 @@ public class LeetCodeText {
             if (i > 0 && nums[i] == nums[i - 1]) {
                 continue;
             }
-            if (nums[i] + nums[i + 1] + nums[i + 2] + nums[i + 3] > target) {
+            if ((long) nums[i] + nums[i + 1] + nums[i + 2] + nums[i + 3] > target) {
                 break;
             }
-            if (nums[i] + nums[nums.length - 1] + nums[nums.length - 2] + nums[nums.length - 3] < target) {
+            if ((long) nums[i] + nums[nums.length - 1] + nums[nums.length - 2] + nums[nums.length - 3] < target) {
                 continue;
             }
             for (int j = i + 1; j < nums.length - 2; ++j) {
                 if (j > i + 1 && nums[j] == nums[j - 1]) {
                     continue;
                 }
-                if (nums[i] + nums[j] + nums[j + 1] + nums[j + 2] > target) {
+                if ((long) nums[i] + nums[j] + nums[j + 1] + nums[j + 2] > target) {
                     break;
                 }
-                if (nums[i] + nums[j] + nums[nums.length - 1] + nums[nums.length - 2] < target) {
+                if ((long) nums[i] + nums[j] + nums[nums.length - 1] + nums[nums.length - 2] < target) {
                     continue;
                 }
                 int left = j + 1;
                 int right = nums.length - 1;
                 while (left < right) {
-                    int sum = nums[i] + nums[j] + nums[left] + nums[right];
+                    long sum = (long) nums[i] + nums[j] + nums[left] + nums[right];
                     if (sum == target) {
                         list.add(Arrays.asList(nums[i], nums[j], nums[left], nums[right]));
                         while (left < right && nums[left] == nums[left + 1]) {
